Add '/var/lib/longhorn1' to default disks for carly host.
Note that this is just for the _default_ disks, however. Longhorn only initializes the default disks on initial bring-up
of a node. After that, when you add additional disks, you need to either add them via the web UI or via updating
_I think_ the `nodes.longhorn.io` API object for the given node.

So changing this host variable is essentially useful for documentation, and maybe as a nice-to-have for if my rootfs
disk dies (though not sure if there will be other steps in that latter case to protect against data destruction).